[CAL-770] add new integration architecture revamp
