+++ ee/pages/api/integrations/stripepayment/webhook.ts
@@ -4,11 +4,11 @@ import Stripe from "stripe";
 
 import stripe from "@ee/lib/stripe/server";
 
-import { CalendarEvent } from "@lib/calendarClient";
 import { IS_PRODUCTION } from "@lib/config/constants";
 import { HttpError as HttpCode } from "@lib/core/http/error";
 import { getErrorFromUnknown } from "@lib/errors";
 import EventManager from "@lib/events/EventManager";
+import { CalendarEvent } from "@lib/integrations/calendar/interfaces/Calendar";
 import prisma from "@lib/prisma";
 import { Ensure } from "@lib/types/utils";
