+++ lib/integrations/calendar/constants/defaults.ts
@@ -1,0 +1,11 @@
+export const DEFAULT_CALENDAR_INTEGRATION_NAME = "";
+
+export const DEFAULT_PASSWORD = "";
+
+export const DEFAULT_URL = "";
+
+export const DEFAULT_CALENDAR_DISPLAY_NAME = "";
+
+export const DEFAULT_TIMEZONE = "";
+
+export const DEFAULT_BOOKING_UID = "";
