+++ lib/integrations/calendar/constants/formats.ts
@@ -1,0 +1,1 @@
+export const TIMEZONE_FORMAT = "YYYY-MM-DDTHH:mm:ss[Z]";
