+++ lib/integrations/calendar/constants/generals.ts
@@ -1,0 +1,19 @@
+export const VTIMEZONE_CALENDAR_COMPONENT = "vtimezone";
+export const VEVENT_CALENDAR_COMPONENT = "VEVENT";
+
+export const TZID_CALENDAR_PROPERTY = "tzid";
+
+export const CALDAV_CALENDAR_TYPE = "caldav";
+
+export const ATTENDEE_NEEDS_ACTION = "NEEDS-ACTION";
+
+export const UTC_TIMEZONE_TYPE = "utc";
+
+export const APPLE_CALENDAR_URL = "https://caldav.icloud.com";
+
+export const CALENDAR_INTEGRATIONS_TYPES = {
+  apple: "apple_calendar",
+  caldav: "caldav_calendar",
+  google: "google_calendar",
+  office365: "office365_calendar",
+};
