+++ lib/integrations/calendar/interfaces/GoogleCalendar.ts
@@ -1,0 +1,5 @@
+import { calendar_v3 } from "googleapis";
+
+export interface ConferenceData {
+  createRequest?: calendar_v3.Schema$CreateConferenceRequest;
+}
