+++ lib/integrations/calendar/interfaces/Office365Calendar.ts
@@ -1,0 +1,10 @@
+export type BufferedBusyTime = {
+  start: string;
+  end: string;
+};
+
+export type O365AuthCredentials = {
+  expiry_date: number;
+  access_token: string;
+  refresh_token: string;
+};
