+++ lib/integrations/calendar/services/BaseCalendarService.ts
@@ -1,11 +1,11 @@
 import { Credential } from "@prisma/client";
 import dayjs from "dayjs";
-import utc from "dayjs/plugin/utc";
 import ICAL from "ical.js";
-import { Attendee, createEvent, DateArray, DurationObject } from "ics";
+import { createEvent } from "ics";
 import {
   createAccount,
   createCalendarObject,
+  DAVAccount,
   deleteCalendarObject,
   fetchCalendarObjects,
   fetchCalendars,
@@ -18,73 +18,73 @@ import { getLocation, getRichDescription } from "@lib/CalEventParser";
 import { symmetricDecrypt } from "@lib/crypto";
 import logger from "@lib/logger";
 
-import { CalendarEvent, IntegrationCalendar } from "./calendarClient";
+import {
+  DEFAULT_CALENDAR_INTEGRATION_NAME,
+  DEFAULT_PASSWORD,
+  DEFAULT_URL,
+  DEFAULT_CALENDAR_DISPLAY_NAME,
+  DEFAULT_TIMEZONE,
+} from "../constants/defaults";
+import { TIMEZONE_FORMAT } from "../constants/formats";
+import {
+  CALDAV_CALENDAR_TYPE,
+  TZID_CALENDAR_PROPERTY,
+  UTC_TIMEZONE_TYPE,
+  VEVENT_CALENDAR_COMPONENT,
+  VTIMEZONE_CALENDAR_COMPONENT,
+} from "../constants/generals";
+import { CalendarEventType, EventBusyDate, NewCalendarEventType } from "../constants/types";
+import { Calendar, CalendarEvent, IntegrationCalendar } from "../interfaces/Calendar";
+import { convertDate, getAttendees, getDuration } from "../utils/CalendarUtils";
 
-dayjs.extend(utc);
+const { CALENDSO_ENCRYPTION_KEY = "" } = process.env;
 
-export type Person = { name: string; email: string; timeZone: string };
-
-export class BaseCalendarApiAdapter {
+export default abstract class BaseCalendarService implements Calendar {
   private url: string;
   private credentials: Record<string, string>;
   private headers: Record<string, string>;
-  private integrationName = "";
+  protected integrationName = DEFAULT_CALENDAR_INTEGRATION_NAME;
+
+  log = logger.getChildLogger({ prefix: [`[[lib] ${this.integrationName}`] });
 
   constructor(credential: Credential, integrationName: string, url?: string) {
-    const decryptedCredential = JSON.parse(
-      symmetricDecrypt(credential.key as string, process.env.CALENDSO_ENCRYPTION_KEY!)
-    );
-    const username = decryptedCredential.username;
-    const password = decryptedCredential.password;
-    this.url = url || decryptedCredential.url;
+    const {
+      username,
+      password,
+      url: credentialURL,
+    } = JSON.parse(symmetricDecrypt(credential.key as string, CALENDSO_ENCRYPTION_KEY));
+
+    this.url = url || credentialURL;
     this.integrationName = integrationName;
     this.credentials = { username, password };
     this.headers = getBasicAuthHeaders({ username, password });
   }
 
-  log = logger.getChildLogger({ prefix: [`[[lib] ${this.integrationName}`] });
-
-  convertDate(date: string): DateArray {
-    return dayjs(date)
-      .utc()
-      .toArray()
-      .slice(0, 6)
-      .map((v, i) => (i === 1 ? v + 1 : v)) as DateArray;
-  }
-
-  getDuration(start: string, end: string): DurationObject {
-    return {
-      minutes: dayjs(end).diff(dayjs(start), "minute"),
-    };
-  }
-
-  getAttendees(attendees: Person[]): Attendee[] {
-    return attendees.map(({ email, name }) => ({ name, email, partstat: "NEEDS-ACTION" }));
-  }
-
-  async createEvent(event: CalendarEvent) {
+  async createEvent(event: CalendarEvent): Promise<NewCalendarEventType> {
     try {
       const calendars = await this.listCalendars(event);
+
       const uid = uuidv4();
-      /** We create local ICS files */
+
+      // We create local ICS files
       const { error, value: iCalString } = createEvent({
         uid,
         startInputType: "utc",
-        start: this.convertDate(event.startTime),
-        duration: this.getDuration(event.startTime, event.endTime),
+        start: convertDate(event.startTime),
+        duration: getDuration(event.startTime, event.endTime),
         title: event.title,
         description: getRichDescription(event),
         location: getLocation(event),
         organizer: { email: event.organizer.email, name: event.organizer.name },
-        // according to https://datatracker.ietf.org/doc/html/rfc2446#section-3.2.1, in a published iCalendar component. "Attendees" MUST NOT be present
-        // attendees: this.getAttendees(event.attendees),
+        /** according to https://datatracker.ietf.org/doc/html/rfc2446#section-3.2.1, in a published iCalendar component.
+         * "Attendees" MUST NOT be present
+         * `attendees: this.getAttendees(event.attendees),`
+         */
       });
 
-      if (error) throw new Error("Error creating iCalString");
+      if (error || !iCalString) throw new Error("Error creating iCalString");
 
-      if (!iCalString) throw new Error("Error creating iCalString");
-
-      /** We create the event directly on iCal */
+      // We create the event directly on iCal
       const responses = await Promise.all(
         calendars
           .filter((c) =>
@@ -98,7 +98,8 @@ export class BaseCalendarApiAdapter {
                 url: calendar.externalId,
               },
               filename: `${uid}.ics`,
-              // according to https://datatracker.ietf.org/doc/html/rfc4791#section-4.1, Calendar object resources contained in calendar collections MUST NOT specify the iCalendar METHOD property.
+              /** according to https://datatracker.ietf.org/doc/html/rfc4791#section-4.1,
+               * Calendar object resources contained in calendar collections MUST NOT specify the iCalendar METHOD property. */
               iCalString: iCalString.replace(/METHOD:[^\r\n]+\r\n/g, ""),
               headers: this.headers,
             })
@@ -115,48 +116,42 @@ export class BaseCalendarApiAdapter {
         uid,
         id: uid,
         type: this.integrationName,
-        password: "",
-        url: "",
+        password: DEFAULT_PASSWORD,
+        url: DEFAULT_URL,
+        additionalInfo: {},
       };
     } catch (reason) {
-      console.error(reason);
+      logger.error(reason);
+
       throw reason;
     }
   }
 
-  async updateEvent(uid: string, event: CalendarEvent) {
+  async updateEvent(uid: string, event: CalendarEvent): Promise<any> {
     try {
-      const calendars = await this.listCalendars();
-      const events = [];
-
-      for (const cal of calendars) {
-        const calEvents = await this.getEvents(cal.externalId, null, null, [`${cal.externalId}${uid}.ics`]);
-
-        for (const ev of calEvents) {
-          events.push(ev);
-        }
-      }
+      const events = await this.getEventsByUID(uid);
 
       const { error, value: iCalString } = createEvent({
         uid,
-        startInputType: "utc",
-        start: this.convertDate(event.startTime),
-        duration: this.getDuration(event.startTime, event.endTime),
+        startInputType: UTC_TIMEZONE_TYPE,
+        start: convertDate(event.startTime),
+        duration: getDuration(event.startTime, event.endTime),
         title: event.title,
         description: getRichDescription(event),
         location: getLocation(event),
         organizer: { email: event.organizer.email, name: event.organizer.name },
-        attendees: this.getAttendees(event.attendees),
+        attendees: getAttendees(event.attendees),
       });
 
       if (error) {
         this.log.debug("Error creating iCalString");
+
         return {};
       }
 
       const eventsToUpdate = events.filter((event) => event.uid === uid);
 
-      return await Promise.all(
+      return Promise.all(
         eventsToUpdate.map((event) => {
           return updateCalendarObject({
             calendarObject: {
@@ -169,28 +164,20 @@ export class BaseCalendarApiAdapter {
         })
       );
     } catch (reason) {
-      console.error(reason);
+      this.log.error(reason);
+
       throw reason;
     }
   }
 
-  async deleteEvent(uid: string): Promise<void> {
+  async deleteEvent(uid: string): Promise<any> {
     try {
-      const calendars = await this.listCalendars();
-      const events = [];
+      const events = await this.getEventsByUID(uid);
 
-      for (const cal of calendars) {
-        const calEvents = await this.getEvents(cal.externalId, null, null, [`${cal.externalId}${uid}.ics`]);
-
-        for (const ev of calEvents) {
-          events.push(ev);
-        }
-      }
-
-      const eventsToUpdate = events.filter((event) => event.uid === uid);
+      const eventsToDelete = events.filter((event) => event.uid === uid);
 
       await Promise.all(
-        eventsToUpdate.map((event) => {
+        eventsToDelete.map((event) => {
           return deleteCalendarObject({
             calendarObject: {
               url: event.url,
@@ -201,62 +188,37 @@ export class BaseCalendarApiAdapter {
         })
       );
     } catch (reason) {
-      console.error(reason);
+      this.log.error(reason);
+
       throw reason;
     }
   }
 
-  async getAvailability(dateFrom: string, dateTo: string, selectedCalendars: IntegrationCalendar[]) {
-    try {
-      const selectedCalendarIds = selectedCalendars
-        .filter((e) => e.integration === this.integrationName)
-        .map((e) => e.externalId);
-      if (selectedCalendarIds.length == 0 && selectedCalendars.length > 0) {
-        // Only calendars of other integrations selected
-        return Promise.resolve([]);
-      }
+  getAvailability(
+    dateFrom: string,
+    dateTo: string,
+    selectedCalendars: IntegrationCalendar[]
+  ): Promise<EventBusyDate[]> {
+    this.log.info(`dateFrom: ${dateFrom}, dateTo: ${dateTo}, selectedCalendars: ${selectedCalendars}`);
 
-      return (
-        selectedCalendarIds.length === 0
-          ? this.listCalendars().then((calendars) => calendars.map((calendar) => calendar.externalId))
-          : Promise.resolve(selectedCalendarIds)
-      ).then(async (ids: string[]) => {
-        if (ids.length === 0) {
-          return Promise.resolve([]);
-        }
-
-        return (
-          await Promise.all(
-            ids.map(async (calId) => {
-              return (await this.getEvents(calId, dateFrom, dateTo)).map((event) => {
-                return {
-                  start: event.startDate.toISOString(),
-                  end: event.endDate.toISOString(),
-                };
-              });
-            })
-          )
-        ).flatMap((event) => event);
-      });
-    } catch (reason) {
-      this.log.error(reason);
-      throw reason;
-    }
+    throw new Error("Method not implemented.");
   }
 
   async listCalendars(event?: CalendarEvent): Promise<IntegrationCalendar[]> {
     try {
       const account = await this.getAccount();
+
       const calendars = await fetchCalendars({
         account,
         headers: this.headers,
       });
 
       return calendars.reduce<IntegrationCalendar[]>((newCalendars, calendar) => {
-        if (!calendar.components?.includes("VEVENT")) return newCalendars;
+        if (!calendar.components?.includes(VEVENT_CALENDAR_COMPONENT)) return newCalendars;
+
         newCalendars.push({
           externalId: calendar.url,
-          name: calendar.displayName ?? "",
+          name: calendar.displayName ?? DEFAULT_CALENDAR_DISPLAY_NAME,
           primary: event?.destinationCalendar?.externalId
             ? event.destinationCalendar.externalId === calendar.url
             : false,
@@ -265,12 +227,13 @@ export class BaseCalendarApiAdapter {
         return newCalendars;
       }, []);
     } catch (reason) {
-      console.error(reason);
+      logger.error(reason);
+
       throw reason;
     }
   }
 
-  async getEvents(
+  private async getEvents(
     calId: string,
     dateFrom: string | null,
     dateTo: string | null,
@@ -285,8 +248,8 @@ export class BaseCalendarApiAdapter {
         timeRange:
           dateFrom && dateTo
             ? {
-                start: dayjs(dateFrom).utc().format("YYYY-MM-DDTHH:mm:ss[Z]"),
-                end: dayjs(dateTo).utc().format("YYYY-MM-DDTHH:mm:ss[Z]"),
+                start: dayjs(dateFrom).utc().format(TIMEZONE_FORMAT),
+                end: dayjs(dateTo).utc().format(TIMEZONE_FORMAT),
               }
             : undefined,
         headers: this.headers,
@@ -296,16 +259,21 @@ export class BaseCalendarApiAdapter {
         .filter((e) => !!e.data)
         .map((object) => {
           const jcalData = ICAL.parse(object.data);
+
           const vcalendar = new ICAL.Component(jcalData);
-          const vevent = vcalendar.getFirstSubcomponent("vevent");
+
+          const vevent = vcalendar.getFirstSubcomponent(VEVENT_CALENDAR_COMPONENT.toLowerCase());
           const event = new ICAL.Event(vevent);
 
           const calendarTimezone =
-            vcalendar.getFirstSubcomponent("vtimezone")?.getFirstPropertyValue("tzid") || "";
+            vcalendar
+              .getFirstSubcomponent(VTIMEZONE_CALENDAR_COMPONENT)
+              ?.getFirstPropertyValue(TZID_CALENDAR_PROPERTY) || DEFAULT_TIMEZONE;
 
           const startDate = calendarTimezone
             ? dayjs(event.startDate.toJSDate()).tz(calendarTimezone)
             : new Date(event.startDate.toUnixTime() * 1000);
+
           const endDate = calendarTimezone
             ? dayjs(event.endDate.toJSDate()).tz(calendarTimezone)
             : new Date(event.endDate.toUnixTime() * 1000);
@@ -342,16 +310,30 @@ export class BaseCalendarApiAdapter {
     }
   }
 
-  private async getAccount() {
-    const account = await createAccount({
+  private async getEventsByUID(uid: string): Promise<CalendarEventType[]> {
+    const events = [];
+
+    const calendars = await this.listCalendars();
+
+    for (const cal of calendars) {
+      const calEvents = await this.getEvents(cal.externalId, null, null, [`${cal.externalId}${uid}.ics`]);
+
+      for (const ev of calEvents) {
+        events.push(ev);
+      }
+    }
+
+    return events;
+  }
+
+  private async getAccount(): Promise<DAVAccount> {
+    return createAccount({
       account: {
         serverUrl: this.url,
-        accountType: "caldav",
+        accountType: CALDAV_CALENDAR_TYPE,
         credentials: this.credentials,
       },
       headers: this.headers,
     });
-
-    return account;
   }
 }
