+++ pages/api/book/confirm.ts
@@ -4,10 +4,10 @@ import type { NextApiRequest, NextApiResponse } from "next";
 import { refund } from "@ee/lib/stripe/server";
 
 import { getSession } from "@lib/auth";
-import { CalendarEvent, AdditionInformation } from "@lib/calendarClient";
 import { sendDeclinedEmails } from "@lib/emails/email-manager";
 import { sendScheduledEmails } from "@lib/emails/email-manager";
 import EventManager from "@lib/events/EventManager";
+import { CalendarEvent, AdditionInformation } from "@lib/integrations/calendar/interfaces/Calendar";
 import logger from "@lib/logger";
 import prisma from "@lib/prisma";
 import { BookingConfirmBody } from "@lib/types/booking";
