+++ pages/api/book/event.ts
@@ -11,7 +11,6 @@ import { v5 as uuidv5 } from "uuid";
 
 import { handlePayment } from "@ee/lib/stripe/server";
 
-import { CalendarEvent, AdditionInformation, getBusyCalendarTimes } from "@lib/calendarClient";
 import {
   sendScheduledEmails,
   sendRescheduledEmails,
@@ -21,7 +20,9 @@ import { ensureArray } from "@lib/ensureArray";
 import { getErrorFromUnknown } from "@lib/errors";
 import { getEventName } from "@lib/event";
 import EventManager, { EventResult, PartialReference } from "@lib/events/EventManager";
-import { BufferedBusyTime } from "@lib/integrations/Office365Calendar/Office365CalendarApiAdapter";
+import { getBusyCalendarTimes } from "@lib/integrations/calendar/CalendarManager";
+import { CalendarEvent, AdditionInformation } from "@lib/integrations/calendar/interfaces/Calendar";
+import { BufferedBusyTime } from "@lib/integrations/calendar/interfaces/Office365Calendar";
 import logger from "@lib/logger";
 import notEmpty from "@lib/notEmpty";
 import prisma from "@lib/prisma";
