+++ pages/api/cron/bookingReminder.ts
@@ -2,8 +2,8 @@ import { ReminderType } from "@prisma/client";
 import dayjs from "dayjs";
 import type { NextApiRequest, NextApiResponse } from "next";
 
-import { CalendarEvent } from "@lib/calendarClient";
 import { sendOrganizerRequestReminderEmail } from "@lib/emails/email-manager";
+import { CalendarEvent } from "@lib/integrations/calendar/interfaces/Calendar";
 import prisma from "@lib/prisma";
 
 import { getTranslation } from "@server/lib/i18n";
