+++ pages/api/integrations/caldav/add.ts
@@ -2,7 +2,7 @@ import type { NextApiRequest, NextApiResponse } from "next";
 
 import { getSession } from "@lib/auth";
 import { symmetricEncrypt } from "@lib/crypto";
-import { CalDavCalendar } from "@lib/integrations/CalDav/CalDavCalendarAdapter";
+import { getCalendar } from "@lib/integrations/calendar/CalendarManager";
 import logger from "@lib/logger";
 import prisma from "@lib/prisma";
 
@@ -38,11 +38,11 @@ export default async function handler(req: NextApiRequest, res: NextApiResponse)
     };
 
     try {
-      const dav = new CalDavCalendar({
+      const dav = getCalendar({
         id: 0,
         ...data,
       });
-      await dav.listCalendars();
+      await dav?.listCalendars();
       await prisma.credential.create({
         data,
       });
